+++ src/geo/CMakeLists.txt
@@ -1,3 +1,3 @@
-add_subdirectory(src)
+add_subdirectory(lib)
 add_subdirectory(test)
 add_subdirectory(sample)
